[tina++/crc] typos korrigiert

+++ tina/tina++/crc/crc.h
@@ -22,34 +22,34 @@ namespace CRC8 {
 #if TURAG_CRC_CRC8_ALGORITHM != 0 || defined(__DOXYGEN__)	
 	
 template <typename T> _always_inline
-uint16_t calculate(const T& data) {
+uint8_t calculate(const T& data) {
   return turag_crc8_calculate(std::addressof(data), sizeof(T));
 }
 
 template <typename T, std::size_t N> _always_inline
-uint16_t calculate(const T (&data)[N]) {
+uint8_t calculate(const T (&data)[N]) {
   return turag_crc8_calculate(data, N * sizeof(T));
 }
 
 _always_inline
-uint16_t calculate(const void* data, std::size_t length) {
+uint8_t calculate(const void* data, std::size_t length) {
   return turag_crc8_calculate(data, length);
 }
 
 
 
 template <typename T> _always_inline
-bool check(const T& data, uint16_t chksum) {
+bool check(const T& data, uint8_t chksum) {
   return turag_crc8_check(std::addressof(data), sizeof(T), chksum);
 }
 
 template <typename T, std::size_t N> _always_inline
-bool check(const T (&data)[N], uint16_t chksum) {
+bool check(const T (&data)[N], uint8_t chksum) {
   return turag_crc8_check(data, N * sizeof(T), chksum);
 }
 
 _always_inline
-bool check(const void* data, std::size_t length, uint16_t chksum) {
+bool check(const void* data, std::size_t length, uint8_t chksum) {
   return turag_crc8_check(data, length, chksum);
 }
 
@@ -78,34 +78,34 @@ namespace CRC8_MOW {
 #if TURAG_CRC_CRC8_MOW_ALGORITHM != 0 || defined(__DOXYGEN__)	
 	
 template <typename T> _always_inline
-uint16_t calculate(const T& data) {
+uint8_t calculate(const T& data) {
   return turag_crc8_mow_calculate(std::addressof(data), sizeof(T));
 }
 
 template <typename T, std::size_t N> _always_inline
-uint16_t calculate(const T (&data)[N]) {
+uint8_t calculate(const T (&data)[N]) {
   return turag_crc8_mow_calculate(data, N * sizeof(T));
 }
 
 _always_inline
-uint16_t calculate(const void* data, std::size_t length) {
+uint8_t calculate(const void* data, std::size_t length) {
   return turag_crc8_mow_calculate(data, length);
 }
 
 
 
 template <typename T> _always_inline
-bool check(const T& data, uint16_t chksum) {
+bool check(const T& data, uint8_t chksum) {
   return turag_crc8_mow_check(std::addressof(data), sizeof(T), chksum);
 }
 
 template <typename T, std::size_t N> _always_inline
-bool check(const T (&data)[N], uint16_t chksum) {
+bool check(const T (&data)[N], uint8_t chksum) {
   return turag_crc8_mow_check(data, N * sizeof(T), chksum);
 }
 
 _always_inline
-bool check(const void* data, std::size_t length, uint16_t chksum) {
+bool check(const void* data, std::size_t length, uint8_t chksum) {
   return turag_crc8_mow_check(data, length, chksum);
 }
